+++ src/main/kotlin/cloud/glitchdev/rfu/data/fishing/FishTrackingType.kt
@@ -1,0 +1,8 @@
+package cloud.glitchdev.rfu.data.fishing
+
+enum class FishTrackingType(val displayName: String) {
+    SC_H("SC/h"),
+    XP_H("XP/h"),
+    TIMER("Timer"),
+    OVERALL("Overall")
+}