+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Features:
 
 ### Changes:
+- Merged XP/h and SC/h displays into a single Fish Tracking display.
+- Unified fishing tracking settings.
+- Made the "overall" part of the fishing tracker more compact.
 
 ### Fixes: